requirements.txt
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
google-api-core==1.16.0
google-auth==1.11.2
google-cloud-core==1.3.0
google-cloud-storage==1.26.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
loguru==0.4.1
MarkupSafe==1.1.1
Pillow==7.0.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==1.7.1
PyMySQL==0.9.3
python-dotenv==0.12.0
python-resize-image==1.1.19
pytz==2019.3
requests==2.23.0
rsa==4.0
six==1.14.0
SQLAlchemy==1.3.13
urllib3==1.25.8
webp-converter==4.0.1
Werkzeug==1.0.0
